:art: chore: 优化模型本地存取
