+++ src/constants/dance.ts
@@ -3,15 +3,14 @@ import { Dance } from '@/types/dance';
 export const DEFAULT_DANCE: Dance = {
   danceId: 'vidol-dance-gokuraku',
   name: '極楽浄土',
+  author: 'rdmclin2',
+  homepage: 'https://github.com/rdmclin2',
   createAt: '2023-10-31',
-  author: 'Vidol',
-  homepage: 'https://github.com/lobehub/lobe-vidol',
-  src: 'https://registry.npmmirror.com/@v-idol/vidol-dance-gokuraku/1.0.0/files/gokuraku.vmd',
-  audio:
-    'https://registry.npmmirror.com/@v-idol/vidol-dance-gokuraku/1.0.0/files/Gokuraku jodo.mp3',
-  cover: 'https://registry.npmmirror.com/@v-idol/vidol-dance-gokuraku/1.0.0/files/cover.jpg',
-  thumb: 'https://registry.npmmirror.com/@v-idol/vidol-dance-gokuraku/1.0.0/files/thumb.jpg',
+  src: 'https://r2.vidol.chat/dances/vidol-dance-gokuraku/gokuraku.vmd',
+  audio: 'https://r2.vidol.chat/dances/vidol-dance-gokuraku/Gokuraku jodo.mp3',
+  cover: 'https://r2.vidol.chat/dances/vidol-dance-gokuraku/cover.jpg',
+  thumb: 'https://r2.vidol.chat/dances/vidol-dance-gokuraku/thumb.jpg',
   readme:
-    '﻿WHAT CHANGED?!\r\n-FIXED THE GOOGLY/SHAKY EYES\r\n-Added eye motions that were supposed to be there\r\n-Audio is slightly more in sync\r\n\r\nSong: 【GARNiDELiA】極楽浄土【とく×メイリア】\r\nhttps://www.nicovideo.jp/watch/sm28709142\r\n\r\nMotion by yurie\r\nhttps://www.nicovideo.jp/watch/sm29180863',
+    '\uFEFFWHAT CHANGED?!\r\n-FIXED THE GOOGLY/SHAKY EYES\r\n-Added eye motions that were supposed to be there\r\n-Audio is slightly more in sync\r\n\r\nSong: 【GARNiDELiA】極楽浄土【とく×メイリア】\r\nhttps://www.nicovideo.jp/watch/sm28709142\r\n\r\nMotion by yurie\r\nhttps://www.nicovideo.jp/watch/sm29180863',
   schemaVersion: 1,
 };
