+++ src/store/dance/selectors/dance.ts
@@ -1,6 +1,4 @@
 import { Dance } from '@/types/dance';
-import { getAudioPathByDanceId, getDancePathByDanceId } from '@/utils/file';
-import storage from '@/utils/storage';
 
 import { DanceStore } from '../index';
 
@@ -21,43 +19,8 @@ const subscribed = (s: DanceStore) => (danceId: string) => {
   return index !== -1;
 };
 
-const getCurrentPlayData = async (
-  s: DanceStore,
-): Promise<
-  | {
-      audioBlob: Blob;
-      danceBuffer: ArrayBuffer;
-    }
-  | undefined
-> => {
-  const currentPlay = s.currentPlay;
-  if (!currentPlay) return undefined;
-
-  // 舞蹈文件
-  const localDancePath = getDancePathByDanceId(currentPlay.danceId);
-  let danceBuffer = (await storage.getItem(localDancePath)) as ArrayBuffer;
-  if (!danceBuffer) {
-    danceBuffer = await fetch(currentPlay.src).then((res) => res.arrayBuffer());
-    await storage.setItem(localDancePath, danceBuffer);
-  }
-
-  // 音频文件
-  const localAudioPath = getAudioPathByDanceId(currentPlay.danceId);
-  const audioBlob = (await storage.getItem(localAudioPath)) as Blob;
-  if (!audioBlob) {
-    const audioBlob = await fetch(currentPlay.audio).then((res) => res.blob());
-    await storage.setItem(localAudioPath, audioBlob);
-  }
-
-  return {
-    danceBuffer: danceBuffer,
-    audioBlob: audioBlob,
-  };
-};
-
 export const danceListSelectors = {
   currentDanceItem,
   showSideBar,
-  getCurrentPlayData,
   subscribed,
 };
