+++ src/utils/file.ts
@@ -1,3 +1,6 @@
+import { Dance } from '@/types/dance';
+import storage from '@/utils/storage';
+
 const MODEL_SCHEMA = 'model';
 const AUDIO_SCHEMA = 'audio';
 const Dance_SCHEMA = 'dance';
@@ -12,4 +15,37 @@ export const getAudioPathByDanceId = (id: string) => {
 
 export const getDancePathByDanceId = (id: string) => {
   return `${Dance_SCHEMA}://${id}`;
+};
+
+export const getCurrentPlayData = async (
+  currentPlay: Dance,
+): Promise<
+  | {
+      audioBlob: Blob;
+      danceBuffer: ArrayBuffer;
+    }
+  | undefined
+> => {
+  if (!currentPlay) return undefined;
+
+  // 舞蹈文件
+  const localDancePath = getDancePathByDanceId(currentPlay.danceId);
+  let danceBuffer = (await storage.getItem(localDancePath)) as ArrayBuffer;
+  if (!danceBuffer) {
+    danceBuffer = await fetch(currentPlay.src).then((res) => res.arrayBuffer());
+    await storage.setItem(localDancePath, danceBuffer);
+  }
+
+  // 音频文件
+  const localAudioPath = getAudioPathByDanceId(currentPlay.danceId);
+  const audioBlob = (await storage.getItem(localAudioPath)) as Blob;
+  if (!audioBlob) {
+    const audioBlob = await fetch(currentPlay.audio).then((res) => res.blob());
+    await storage.setItem(localAudioPath, audioBlob);
+  }
+
+  return {
+    danceBuffer: danceBuffer,
+    audioBlob: audioBlob,
+  };
 };
